requirements.txt
# --- Core / Config / CLI ---
pydantic>=2.7
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
rich>=13.7.1
typer>=0.12.3
pendulum>=3.0.0
holidays>=0.57
PyMuPDF>=1.24.0 
python-dateutil>=2.9
pandas>=2.2
camelot-py>=0.11.0 
pdfplumber>=0.11.0
pdfminer.six>=20231228

# --- HTTP / Similarity ---
requests>=2.32.3
rapidfuzz>=3.9.6
httpx>=0.27.0

# Networking layer differences (handle macOS Python 3.9 + LibreSSL cleanly)
urllib3<2; python_version < "3.10"
urllib3>=2.5.0; python_version >= "3.10"

# --- Database (Supabase) ---
supabase>=2.4.0

# --- Email ---
sendgrid>=6.11.0

# --- Optional: only install Selenium on Python 3.10+ (to avoid urllib3 conflict on 3.9) ---
selenium>=4.23.1; python_version >= "3.10"
webdriver-manager>=4.0.2; python_version >= "3.10"
playwright>=1.45
cloudscraper>=1.2.71
brotli
